app/requirements.txt
amqp==5.1.1
asgiref==3.5.2
attrs==21.4.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Deprecated==1.1.5
Django==4.0.6
django-filter==22.1
django-redis-cache==3.0.1
djangorestframework==3.13.1
drf-yasg==1.20.0
environs==9.5.0
fpdf2==2.5.5
idna==3.3
importlib-resources==5.9.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.6.2
kombu==5.2.4
MarkupSafe==2.1.1
marshmallow==3.17.0
packaging==21.3
Pillow==9.2.0
prompt-toolkit==3.0.30
pyparsing==3.0.9
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
redis==3.5.3
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
swagger-spec-validator==2.7.4
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.10
vine==5.0.0
wcwidth==0.2.5
zipp==3.8.1
